✨ feat: 2.0.0.a77
* strategy增加lifecycle
* 保留最后一个回测周期仅供交易使用，不调用`predict`
* Security获取股票列表时，如果不调用`types`，将获取股票列表，调用`types()`不传参数将获取带指数、股票的列表。

+++ omicron/plotting/metrics.py
@@ -62,7 +62,8 @@ class MetricsGraph:
 
         if indicator is not None:
             self.indicator = indicator.join(
-                pd.Series(index=self.frames, name="frames"), how="right"
+                pd.Series(index=self.frames, name="frames", dtype=np.float64),
+                how="right",
             )
         else:
             self.indicator = None
